lpc_periph.v: make the code work on hardware
Clock edges related to LAD handling work differently than I thought,
they had to be reversed. In addition, reading doesn't work when
there is code allowing for aborting the cycle on SYNC in block
responsible for changing `fsm_next_state`. Moving it to complementary
`always` block fixes this problem.

SERIRQ has to be physically disconnected, otherwise it impacts read
values, most likely due to electromagnetic interference. Because of
that it was not tested, neither its clock edges were changed.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>